Explain, Verbose, Profile

+++ queries/monitor_create.sql
@@ -3,9 +3,9 @@ CREATE TABLE IF NOT EXISTS TABLENAME (
         start_timestamp TIMESTAMP,
         transaction_id BIGINT,
         statement_id BIGINT,
-        query_duration_us DECIMAL(20,1),
+        query_duration_us NUMERIC(20,5),
         resource_request_execution_ms BIGINT,
-        used_memory_kb DECIMAL(20,1),
-        CPU_TIME INTEGER,
+        used_memory_kb NUMERIC(20,5),
+        CPU_TIME BIGINT,
         label VARCHAR(100)
         );
